Fix: VAST ad are invisible if content video is play in fullscreen.
Issue:

When playing a video the video goes into full screen but on iOS there is a button to open the video in another fullscreen state.
If the video is open in this state then this view is on top of the ad view.

This means when the video ends it stays on screen but the ad plays in the background until the video is exited from full screen mode.

Solution:

We can not disallow user from entering full screen mode to play the content video, because the full screen toggle as one of the default playback controls which is out of our control.

Changes in this commit are:
1. Track if video player is enter into or quit from full screen.
2. If it is in full screen mode, set 'exitsFullScreenWhenPlaybackEnds' to true to AVPlayerViewController.
3. On AVPlayerItemDidPlayToEndTimeNotification callback, postpone a short while leaving chance to quit AVFullScreenViewController if content video was playing in full screen.